Fix: Internal device ID was set as BLE address
ble_hs_id_set_rnd() has a check to discard address if it is not random
(upper 2 bits should be either 00 or 11 for random addresses).
This resulted in the supplied random number not being set as the device's
BLE address and the chip's internal static device ID was used to form
the BLE address, reveling the devices identity.